Fix signedness
git-svn-id: https://svn.smr.arpa.emr.it/dballe/dballe/trunk@2412 5c06e07b-eeef-0310-a3fe-9f68d50baefc

+++ dballe/bufrex/tests/bufr_encoder.cc
@@ -41,7 +41,7 @@ TESTGRP(bufr_encoder);
 
 bool memfind(dba_rawmsg rmsg, const char* str, size_t len)
 {
-	for (int i = 0; true; ++i)
+	for (size_t i = 0; true; ++i)
 	{
 		if (i + len >= rmsg->len) return false;
 		if (memcmp(rmsg->buf + i, str, len) == 0)
